pyproject.toml
[project]
name = "docmind"
version = "0.1.0"
description = "Scalable RAG application for document processing and AI-powered search"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi>=0.115.13,<0.116.0",
    "uvicorn[standard]>=0.34.3,<0.35.0",
    "pydantic>=2.0.0,<3.0.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "psycopg2-binary>=2.9.0,<3.0.0",
    "qdrant-client>=1.7.0,<2.0.0",
    "openai>=1.0.0,<2.0.0",
    "python-multipart>=0.0.6,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "aiofiles>=23.0.0,<24.0.0",
    "pypdf>=3.17.0,<4.0.0",
    "python-docx>=1.1.0,<2.0.0",
    "tiktoken>=0.5.0,<1.0.0",
    "numpy>=1.26.0,<2.0.0",
    "tenacity (>=9.1.2,<10.0.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)"
]

[tool.poetry]
packages = [
    { include = "docmind" }
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^25.1.0"
flake8 = "^7.2.0"
mypy = "^1.8.0"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=docmind",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml"
]
asyncio_mode = "auto"
